# Use CreateNumpyVector in ev42 serialisation