Contributes to ECDC-2191

Performance change:

| Number of events | Old implementation (seconds) | New implementation (seconds) |
|------------------|------------------------------|------------------------------|
| 10000            | 0.087                        | 0.00022                      |
| 100000           | 0.82                         | 0.0016                       |
| 1000000          | 8.10                         | 0.013                        |