@matthew-d-jones 2020-12-22
```python
def wrapper(func, *args, **kwargs):
    def wrapped():
        return func(*args, **kwargs)

    return wrapped


if __name__ == "__main__":
    events = 1_000_000
    tof = np.random.randint(0, 100_000, events, dtype=np.uint32)
    det_id = np.random.randint(0, 100_000, events, dtype=np.uint32)
    source_name = "source"
    message_id = 1
    pulse_time = 123_456_789

    wrapped = wrapper(serialise_ev42, source_name, message_id, pulse_time, tof, det_id)
    wrapped_old = wrapper(serialise_ev42_old, source_name, message_id, pulse_time, tof, det_id)

    print(timeit.timeit(wrapped, number=1))
    print(timeit.timeit(wrapped_old, number=1))
```